Move create_rpaid_test_statistics into a separate module with separate tests.
